% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betareg_lasso.R
\name{betareg_lasso}
\alias{betareg_lasso}
\title{Fit beta model with L1 penalization}
\usage{
betareg_lasso(X, y, lambda, standarize = TRUE, ...)
}
\arguments{
\item{X}{matrix}

\item{y}{response vector}

\item{lambda}{penalization parameter}

\item{standarize}{logical.}

\item{...}{other arguments}
}
\value{
fit
}
\description{
Fit beta model with L1 penalization
}
