% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.penalizedbeta.R
\name{predict.penalizedbeta}
\alias{predict.penalizedbeta}
\title{Predict}
\usage{
\method{predict}{penalizedbeta}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from "penalizedbeta".}

\item{newdata}{optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted linear predictors are used.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
predictions
}
\description{
Predict
}
