% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betareg_lasso.R
\name{betareg_lasso}
\alias{betareg_lasso}
\title{Fit beta model with L1 penalization}
\usage{
betareg_lasso(X, y, lambda, standarize = TRUE, ...)
}
\arguments{
\item{X}{matrix}

\item{y}{response vector}

\item{lambda}{penalization parameter}

\item{standarize}{logical.}

\item{...}{other arguments}
}
\value{
fit
}
\description{
Fit beta model with L1 penalization
}
\examples{

library(penalizedbeta)

y <- data$y
X <- data[,-1]

n <- dim(X)[1]; p <- dim(X)[2]
split <- sample(x = c("train","test"), size = n,
                prob = c(0.8,0.2), replace = TRUE)
Xtrain <- X[split == "train",] ; Xtest <- X[split == "test",]
ytrain <- y[split == "train"] ; ytest <- y[split == "test"]

nlambda = 100
lambdas <-  seq(0.0001,0.1, length.out = nlambda)
s <- lambdas[1]

fit <- betareg_lasso(Xtrain,ytrain,lambda = s)

ypred_test <- predict(fit, newdata = Xtest)
mean((ytest-ypred_test)^2)
}
